refactored graphql server plugin to npm package
