+++ test-app/server/server.ts
@@ -1,14 +1,14 @@
 import { ApolloServer } from "@apollo/server";
-import { expressMiddleware } from '@apollo/server/express4'
+import { expressMiddleware } from '@apollo/server/express4';
 // import { ApolloServerPlugin } from "apollo-server-plugin-base";
-import express from 'express'
-import cors from 'cors'
-import bodyParser from 'body-parser'
+import express from 'express';
+import cors from 'cors';
+import bodyParser from 'body-parser';
 import { typeDefs } from "./schema";
 import { resolvers } from "./resolvers";
-import testDb from './models/testDb'
+import testDb from './models/testDb';
 // import { GraphQLRequestContext } from "apollo-server-types";
-import { getProjectId, insertMetrics } from 'kensa-api';
+import { getProjectId, insertMetrics, testPlugin } from 'kensa-api';
 
 
 async function startApolloServer() {
@@ -16,80 +16,88 @@ async function startApolloServer() {
   const PORT = process.env.PORT || 3000
 
   // Move plugins to npm package later
-  const pluginOne = {
-    // Fires whenever a GraphQL request is received from a client. This plugin runs after whatever happens in ApolloServer's context
-    async requestDidStart(requestContext: any) {
-      // IntrospectionQuery keeps running in GraphQL server. Use this condition to stop context from logging for IntrospectionQuery
-      if (requestContext.request.operationName === 'IntrospectionQuery') return;
-      // Query string request sent by the client
-      // console.log(`Request started! Query: ${requestContext.request.query}`)
-      // console.log(`Request variable! Variable: ${requestContext.request.variables}`)
+  // const pluginOne = {
+  //   // Fires whenever a GraphQL request is received from a client. This plugin runs after whatever happens in ApolloServer's context
+  //   async requestDidStart(requestContext: any) {
+  //     // IntrospectionQuery keeps running in GraphQL server. Use this condition to stop context from logging for IntrospectionQuery
+  //     if (requestContext.request.operationName === 'IntrospectionQuery') return;
+  //     // Query string request sent by the client
+  //     // console.log(`Request started! Query: ${requestContext.request.query}`)
+  //     // console.log(`Request variable! Variable: ${requestContext.request.variables}`)
 
-      const requestStart = Date.now()
-      let op: string
-      return {
-        async executionDidStart(executionRequestContext: any) {
-          return {
-            willResolveField({source, args, contextValue, info}: any) {
-              const resolverStart = Date.now();
-              return (error: any, result: any) => {
-                const resolverEnd = Date.now();
-                console.log(`Field ${info.parentType.name}.${info.fieldName} took ${resolverEnd - resolverStart}ms`)
+  //     const requestStart = Date.now()
+  //     let op: string;
+  //     let success: boolean = true;
+  //     return {
+  //       async executionDidStart(executionRequestContext: any) {
+  //         return {
+  //           willResolveField({source, args, contextValue, info}: any) {
+  //             const resolverStart = Date.now();
+  //             return (error: any, result: any) => {
+  //               const resolverEnd = Date.now();
+  //               console.log(`Field ${info.parentType.name}.${info.fieldName} took ${resolverEnd - resolverStart}ms`)
 
-                if (error) {
-                  console.log(`It failed with ${error}`)
-                } else {
-                  console.log(`It returned ${result}`)
-                }
-              }
-            }
-          }
-        },
-        async didResolveOperation(context: any) {
-          // context ==> { logger, cache, schema, request, response, contextValue, metrics, overallCachePolicy, queryHash, source, document }
-          op = context.operationName
-        },
-        async willSendResponse(context: any) {
-          // console.log(context.response.body.singleResult) // response data
-          const receiveResponse = Date.now()
-          const elapsed = receiveResponse - requestStart
-          console.log(`operation=${op} duration=${elapsed}ms`)
+  //               if (error) {
+  //                 console.log(`It failed with ${error}`)
+  //               } else {
+  //                 console.log(`It returned ${result}`)
+  //               }
+  //             }
+  //           }
+  //         }
+  //       },
+  //       async didResolveOperation(context: any) {
+  //         // context ==> { logger, cache, schema, request, response, contextValue, metrics, overallCachePolicy, queryHash, source, document }
+  //         op = context.operationName
+  //       },
+  //       async didEncounterErrors() {
+  //         const receiveResponse = Date.now();
+  //         const elapsed = receiveResponse - requestStart;
+  //         console.log('error took:', elapsed, 'ms');
+  //       },
+  //       async willSendResponse(context: any) {
+  //         // console.log(context.response.body.singleResult) // response data
+  //         const receiveResponse = Date.now()
+  //         const elapsed = receiveResponse - requestStart;
+  //         console.log(`operation=${op} duration=${elapsed}ms`)
 
-          // Getting projectId from context object
-          const { id } = context.contextValue.projectId;
-          const query_string = context.request.query;
-          // construct a fake response for now
-          const fakeResponse = {
-            query_string: query_string,
-            execution_time: elapsed,
-            success: true,
-            operation_name: op
-          }
-          // This insert metrics to Kensa database
-          await insertMetrics(fakeResponse, id);
-        }
-      }
-    }
-  }
+  //         // Getting projectId from context object
+  //         const { id } = context.contextValue.projectId;
+  //         const query_string = context.request.query;
+  //         // construct an object with metrics
+  //         const metricBody = {
+  //           query_string: query_string,
+  //           execution_time: elapsed,
+  //           success: success,
+  //           operation_name: op
+  //         }
+  //         // This insert metrics to Kensa database
+  //         await insertMetrics(metricBody, id);
+  //       }
+  //     }
+  //   }
+  // }
 
   const apolloServer = new ApolloServer({
     typeDefs,
     resolvers,
-    plugins: [pluginOne]
+    plugins: [testPlugin]
+    // plugins: [pluginOne]
   })
 
   await apolloServer.start();
   
   app.use('/graphql', cors(), bodyParser.json(), expressMiddleware(apolloServer, {
+    // move the context async function to npm package, let user pass in api key and testDb
     context: async ({req, res}: any) => {
       // IntrospectionQuery keeps running. Use this to stop context from logging for IntrospectionQuery
       if (req.body.operationName === 'IntrospectionQuery') return;
       console.log('inside context ApolloServer test-app')
-      console.log(req.body)
+      console.log(req.body);
       // insert your apiKey here
       const apiKey = 'bcdf2ec0-47ae-400b-8aeb-a4a78daf2f05';
       // Calling npm package to get projectId
-      const projectId = await getProjectId(apiKey)
+      const projectId = await getProjectId(apiKey);
       return {
         req,
         res,
@@ -97,7 +105,7 @@ async function startApolloServer() {
         projectId
       }
     },
-  }))
+  }));
   
   app.listen(PORT, () => console.log(`🚀 Server running on port ${PORT}`))
